[ANDROSDK-2017] Move DataApprovalService and Fields to network package
